refactor(ui): Extract create run form's logic to helper functions
The create run form component has gotten quite large and hard to follow.
Simplify it a little by extracting the logic to functions, for which tests
can be written in a follow-up.
